+++ src/model/guild/mod.rs
@@ -2395,7 +2395,7 @@ impl Guild {
     /// [`utils::shard_id`]: crate::utils::shard_id
     #[cfg(all(feature = "cache", feature = "utils"))]
     #[inline]
-    pub fn shard_id(&self, cache: impl AsRef<Cache>) -> u64 {
+    pub fn shard_id(&self, cache: impl AsRef<Cache>) -> u32 {
         self.id.shard_id(&cache)
     }
 
@@ -2422,7 +2422,7 @@ impl Guild {
     #[cfg(all(feature = "utils", not(feature = "cache")))]
     #[inline]
     #[must_use]
-    pub fn shard_id(&self, shard_count: u64) -> u64 {
+    pub fn shard_id(&self, shard_count: u32) -> u32 {
         self.id.shard_id(shard_count)
     }
